Move thread pool logic into separate file
