+++ src/lib.rs
@@ -5,12 +5,9 @@ mod app;
 pub use app::MacroSolverApp;
 
 mod config;
+mod thread_pool;
 mod widgets;
 
 #[cfg(target_arch = "wasm32")]
 pub static OOM_PANIC_OCCURED: std::sync::atomic::AtomicBool =
-    std::sync::atomic::AtomicBool::new(false);
-
-#[cfg(target_arch = "wasm32")]
-pub static THREAD_POOL_IS_INITIALIZED: std::sync::atomic::AtomicBool =
     std::sync::atomic::AtomicBool::new(false);
